backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.32
pydantic==2.9.2
python-dotenv==1.0.1
bcrypt==4.2.0
requests==2.32.3